LC-palindrome-number [Runtime: 8 ms; Memory: 8.6 MB]

+++ math/palindrome-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         int temp = x;
-        int num = 0;
+        long num = 0;
         while(temp>0){
             num = num*10 + temp % 10;
             temp = temp / 10;
